31º Envio Revisao controle de seçoes

+++ app/Http/Controllers/VendaController.php
@@ -20,27 +20,38 @@ class VendaController extends Controller
 	}
 	
 	function telaAdicionarItem($id){
-		$venda = Venda::find($id);
-		$produto = Produto::all();
+		if (session()->has("login")){
+			$venda = Venda::find($id);
+			$produto = Produto::all();
 
-		return view('telas_cadastro.cadastro_itens')->with(compact('venda','produto'));
+			return view('telas_cadastro.cadastro_itens')->with(compact('venda','produto'));
+			
+		}else{
+		return view('tela_login');
+		}
 		
 	}
 	
     function adicionar(Request $req){
-		$id_usuario = $req->input('id_usuario');
-    	
-		$vnd = new Venda();
-    	$vnd->id_usuario = $id_usuario;
-    	$vnd->valor = 0;
-    	
+		if (session()->has("login")){
+			$id_usuario = $req->input('id_usuario');
+			
+			$vnd = new Venda();
+			$vnd->id_usuario = $id_usuario;
+			$vnd->valor = 0;
+			
 
-    	if ($vnd->save()){
-			echo  "<script>alert('Venda efetuada com Sucesso!');</script>";
-    	} else {
-    		echo  "<script>alert('Venda nao efetuada!');</script>";
-		}
-		return redirect()->route('vendas_item_novo', ['id' => $vnd->id]);
+			if ($vnd->save()){
+				echo  "<script>alert('Venda efetuada com Sucesso!');</script>";
+			} else {
+				echo  "<script>alert('Venda nao efetuada!');</script>";
+			}
+			return redirect()->route('vendas_item_novo', ['id' => $vnd->id]);
+
+		}else{
+            return view('tela_login');
+        }
+
 	}
 	
 	function excluir($id){
@@ -103,59 +114,74 @@ class VendaController extends Controller
 	}
 
 	function itensVenda($id){
-		$venda = Venda::find($id);
-		return view('listas.lista_itens_venda', ['venda' => $venda]);
-
+		if (session()->has("login")){
+			$venda = Venda::find($id);
+			return view('listas.lista_itens_venda', ['venda' => $venda]);
+		}
+		return view('tela_login');
 	}
 
 	function adicionarItem(Request $req, $id){
-		$id_produto = $req->input('id_produto');
-		$qtd = $req->input('quantidade');
-		
-		$quantidade = intval($qtd);
+		if (session()->has("login")){
+			$id_produto = $req->input('id_produto');
+			$qtd = $req->input('quantidade');
+			
+			$quantidade = intval($qtd);
 
-		$produto = Produto::find($id_produto);
-		$venda = Venda::find($id);
-		$subtotal = $produto->preco * $quantidade;
+			$produto = Produto::find($id_produto);
+			$venda = Venda::find($id);
+			$subtotal = $produto->preco * $quantidade;
 
-		$colunas_pivot = [
-				'quantidade' => $quantidade,
-				'subtotal' => $subtotal
-		];
+			$colunas_pivot = [
+					'quantidade' => $quantidade,
+					'subtotal' => $subtotal
+			];
 
-		
-		$venda->produtos()->attach($produto->id, $colunas_pivot);
-		$venda->valor += $subtotal;
-		$venda->save();		
-		return redirect()->route('vendas_item_novo', ['id' => $venda->id]);
+			
+			$venda->produtos()->attach($produto->id, $colunas_pivot);
+			$venda->valor += $subtotal;
+			$venda->save();		
+			return redirect()->route('vendas_item_novo', ['id' => $venda->id]);
+		}else{
+			return view('tela_login');
+		}
 	}
 
 	function excluirItem($id , $id_pivot){
-		$venda = Venda::find($id);
-		$subtotal = DB::table('produtos_venda')->where('id',$id_pivot)->value('subtotal');
+		if (session()->has("login")){
+			$venda = Venda::find($id);
+			$subtotal = DB::table('produtos_venda')->where('id',$id_pivot)->value('subtotal');
 
-		$venda->valor = $venda->valor - $subtotal;
-		$venda->produtos()->wherePivot('id','=',$id_pivot)->detach();
-		$venda->save();
+			$venda->valor = $venda->valor - $subtotal;
+			$venda->produtos()->wherePivot('id','=',$id_pivot)->detach();
+			$venda->save();
 
-		return redirect()->route('vendas_item_novo', ['id' => $venda->id]);
+			return redirect()->route('vendas_item_novo', ['id' => $venda->id]);
+		}else{
+			return view('tela_login');
+		}
 
 	}
 
 	function excluirItemLista($id , $id_pivot){
-		$venda = Venda::find($id);
-		$subtotal = DB::table('produtos_venda')->where('id',$id_pivot)->value('subtotal');
+		if (session()->has("login")){
+			$venda = Venda::find($id);
+			$subtotal = DB::table('produtos_venda')->where('id',$id_pivot)->value('subtotal');
 
-		$venda->valor = $venda->valor - $subtotal;
-		$venda->produtos()->wherePivot('id','=',$id_pivot)->detach();
-		$venda->save();
-		$var = DB::table('produtos_venda')->where('id_venda','=',$id)->first();
-		if($var){
-			return view('listas.lista_itens_venda', ['venda' => $venda]);
+			$venda->valor = $venda->valor - $subtotal;
+			$venda->produtos()->wherePivot('id','=',$id_pivot)->detach();
+			$venda->save();
+			$var = DB::table('produtos_venda')->where('id_venda','=',$id)->first();
+			if($var){
+				return view('listas.lista_itens_venda', ['venda' => $venda]);
+			}else{
+				$venda->delete();
+				return redirect()->route('vendas_total');
+			}
 		}else{
-			$venda->delete();
-			return redirect()->route('vendas_total');
+			return view('tela_login');
 		}
+
 	}
 
 	function validar($id){
@@ -167,8 +193,9 @@ class VendaController extends Controller
 				echo "Nao pode adicionar venda sem itens!"; 
 				return redirect()->route('vendas_item_novo', ['id' => $venda->id]);
 			}
-		}
+		}else{
 		return view('tela_login');
+		}
 	}
 
 }
